+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v0.4.0 - TBD
 
++ Bump `LiteDB` version to `5.0.21`
++ Support concurrent operations on the same vault file using LiteDB's shared concurrency model
+
 ## v0.3.0 - 2023-11-22
 
 + Add support for `-WebCredential` when specifying a DPAPI-NG protection descriptor